# src/oas/oas.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"uuid": {
				"type": "string",
				"format": "uuid",
				"description": "Stringified UUIDv4.\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
				"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
			},
			"integer": {
				"type": "integer",
				"format": "int32",
				"description": "A `number` type that resolves as an integer in OpenApi"
			},
			"ConferenceResponse": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/uuid"
					},
					"sourceId": {
						"$ref": "#/components/schemas/integer"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"sourceId",
					"name",
					"active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConferencePostBody": {
				"properties": {
					"sourceId": {
						"$ref": "#/components/schemas/integer"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					}
				},
				"required": [
					"sourceId",
					"name",
					"active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConferencePatchBody": {
				"properties": {
					"sourceId": {
						"$ref": "#/components/schemas/integer"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					}
				},
				"required": [
					"sourceId",
					"name",
					"active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DivisionResponse": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/uuid"
					},
					"sourceId": {
						"$ref": "#/components/schemas/integer"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"conferenceId": {
						"$ref": "#/components/schemas/uuid"
					}
				},
				"required": [
					"id",
					"sourceId",
					"name",
					"active",
					"conferenceId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DivisionPostBody": {
				"properties": {
					"sourceId": {
						"$ref": "#/components/schemas/integer"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"conferenceId": {
						"$ref": "#/components/schemas/uuid"
					}
				},
				"required": [
					"sourceId",
					"name",
					"active",
					"conferenceId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DivisionPatchBody": {
				"properties": {
					"sourceId": {
						"$ref": "#/components/schemas/integer"
					},
					"name": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"conferenceId": {
						"$ref": "#/components/schemas/uuid"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "open-hockey",
		"version": "0.0.1",
		"description": "Awesome project developed with TypeORM.",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/conferences/{id}": {
			"get": {
				"operationId": "GetConference",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConferenceResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Conference"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/uuid"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteConference",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Conference"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/uuid"
						}
					}
				]
			}
		},
		"/conferences": {
			"get": {
				"operationId": "ListConferences",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ConferenceResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conference"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateConference",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConferenceResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Conference"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConferencePostBody"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateConference",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConferenceResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Conference"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConferencePatchBody"
							}
						}
					}
				}
			}
		},
		"/divisions/{id}": {
			"get": {
				"operationId": "GetDivision",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DivisionResponse"
								}
							}
						}
					}
				},
				"description": "Returns a `division` matching the provided `id`",
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/uuid"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateDivision",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DivisionResponse"
								}
							}
						}
					}
				},
				"description": "Updates an existing `division` record",
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/uuid"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DivisionPatchBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteDivision",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a `division` matching the provided `id`",
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/uuid"
						}
					}
				]
			}
		},
		"/divisions": {
			"get": {
				"operationId": "ListDivisions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DivisionResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Returns a list of `divisions` matching the provided query parameters",
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/integer"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/integer"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateDivision",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DivisionResponse"
								}
							}
						}
					}
				},
				"description": "Create a new `division` record",
				"tags": [
					"Division"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DivisionPostBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000/",
			"description": "Local development"
		}
	]
}